pyproject.toml
[project]
name = "stephenkingbooks"
version = "0.1.0"
description = "Agent IA pour la gestion de la bibliographie Stephen King"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "google-genai>=1.52.0",
    "gspread>=6.2.1",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
]
